pyproject.toml
[tool.poetry]
name = "combinestrategy"
version = "0.1.0"
description = ""
authors = ["Alexey <alexeyvasin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
tinkoff-investments = "^0.2.0b108"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
